+++ website/docusaurus.config.ts
@@ -170,6 +170,13 @@ const config: Config = {
         },
 
         {
+          type: "docSidebar",
+          sidebarId: "migrationSidebar",
+          position: "left",
+          label: "Migration",
+        },
+
+        {
           href: "https://github.com/electron-userland/electron-builder",
           label: "GitHub",
           position: "right",